Cleaned up the code little bit
I'm seeing you have your code wrapped around a doOnce clause in OnBarUpdate. With this setup, the code will only run on the first bar of the chart. OnBarUpdate expects to be called for every bar, so all of your calculated values must be in terms of an individual bar. For example, the SMA indicator calculates its moving average for Value[0] always because the historical run fills in the data for you. 

Is a hint I recieved